refactor: 앱 출력 로직 분리
